Added functionality to record a follow me mission. Includes a checkbox setting.
Added "launch" button to flight data screen. Launch spins up the props
for 1 second to overcome the limitation in ardupilot that requires rc
input to begin a mission. I had to change the view to a relitave layout
instead of a linear layout to get the button to appear on top of the
map.

Modified waypoints so that the first waypoint always redefines your home
location.

Added onpause, onresume, and ondestroy methods to rc mode. There was an
error where when you switched to rc mode more than once after the app
was started it would not work. These methods resolved the problem.

Changed the screen to always be in landscape mode for most of the
activities. I was noticing some force close errors intermittently when
switching orientations or screen timeout. I suspect this may have to do
with onpause and on resume and ondestroy not being included in the
flight data mode or it could be because of the layout change I made.