WIP:Breaks the build. Changing the structure of the Mission items.
Instead of reeling on the mavlink message for mission items we now have our own structure. This decouple the app from ArduPilot, making it more resistance to changes on that code.

There are lot's of waypoint types that still need to be created, and all the mission dialogues are broken.