UDP (and possibly others?) would leak connections sometimes
Caused failure to connect the second time you tried to connect via UDP.

Moved the closeConnection into a finally block so it is always invoked.
Possibly a good idea to check for similar places where there are
'close type' operations - those ops should almost always be in a finally
block to prevent leaks.