Clear Guided Waypoint prior to sending mission
During follow-me, DP sends guided waypoints to the copter as the GCS device's location updates.
After finishing follow-me, a waypoint mission always results in the copter pointing in the
direction it was pointing just before stopping the follow-me session. This happens for the entirety
of the mission.

This problem is fixed by sending a MAVLINK_MS_ID_MISSION_CLEAR_ALL (45) message
to the copter immediately before sending the mission.

To test, I loaded a version of DP without this addition on an Android device, and a version
WITH this addition on another Android device. I used the first device, and reproduced the issue
twice. I performed the same test with the second Android device 5 times, and did not observe
the issue.

Fixes #1214