Introducing the following Mapbox API, replaces the original: https://api.mapbox.com/styles/v1/
Original API was depending in the phased out approach with Mapbox projects, and the use of the retured Mapbox Classic app.
The API introduced utilizes the current Mapbox Studio's abilities to create styles, and to modify these styles with tilesets and other customizations
The API requires:
- The Mapbox Account name: As introduced in multiple file as mapboxUserId
- The Style ID: Re-used the mapId for this purpose since the old API is phased out anyway
- The Mapbox token, already present in previous versions of Tower

Solution fully working, but has a flaw when user selects "flight data screen" (no flaw in screens for editor and history:
- Tower loads normal maps from Google, but does not overlay the styled tiles
- Workaround: Change orientation of device, and the styled tiles are overlaid
The flaw is caused by logic in DroneMap.java, the following IF concludes wether or not tiles should be overlaid:
if (mMapFragment == null || mMapFragment.getProvider() != mapProvider) (line 294)

Should be corrected, but Mapbox in Tower is now fully functional with the workaround.

Signed-off-by: oaamaas <olav.aamaas@outlook.com>